+++ CHANGELOG.md
@@ -5,3 +5,10 @@
 ### Changed
 
 - Initial Release
+
+## v1.1.0
+
+### Changed
+
+- Implement createBloombergExcelFile() and createThomsonExcelFile()
+- add testcase test_nav_handler.py